with use of store enhancers we can change how the store API behaves by wrapping around the original Redux stor.
redux enhancer is a funtion that takes createStore as argument and
returns a function with
1. root reudcer
2. preloadedState
3. enhancer
as argument. and creates the store and returns {...store, other properties}
the other properties can be dispach, getState, subscribe (the original
properties of store object's API); in other words we are chnaging the
behaviour of original store.